optimize undo log manager and fix delete undolog support oracle (#1618)
