+++ rm-datasource/src/main/java/io/seata/rm/RMHandlerAT.java
@@ -25,7 +25,7 @@ import io.seata.core.model.ResourceManager;
 import io.seata.core.protocol.transaction.UndoLogDeleteRequest;
 import io.seata.rm.datasource.DataSourceManager;
 import io.seata.rm.datasource.DataSourceProxy;
-import io.seata.rm.datasource.undo.UndoLogManager;
+import io.seata.rm.datasource.undo.UndoLogManagerFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -55,8 +55,8 @@ public class RMHandlerAT extends AbstractRMHandler {
             int deleteRows = 0;
             do {
                 try {
-                    deleteRows = UndoLogManager.deleteUndoLogByLogCreated(logCreatedSave, dataSourceProxy.getDbType(),
-                        LIMIT_ROWS, conn);
+                    deleteRows = UndoLogManagerFactory.getUndoLogManager(dataSourceProxy.getDbType())
+                            .deleteUndoLogByLogCreated(logCreatedSave, LIMIT_ROWS, conn);
                     if (deleteRows > 0 && !conn.getAutoCommit()) {
                         conn.commit();
                     }
