+++ rm-datasource/src/main/java/io/seata/rm/datasource/DataCompareUtils.java
@@ -22,7 +22,7 @@ import io.seata.rm.datasource.sql.struct.Field;
 import io.seata.rm.datasource.sql.struct.Row;
 import io.seata.rm.datasource.sql.struct.TableMeta;
 import io.seata.rm.datasource.sql.struct.TableRecords;
-import io.seata.rm.datasource.undo.UndoLogManager;
+import io.seata.rm.datasource.undo.AbstractUndoLogManager;
 import io.seata.rm.datasource.undo.parser.FastjsonUndoLogParser;
 
 import java.math.BigDecimal;
@@ -62,7 +62,7 @@ public class DataCompareUtils {
                         if (f1.getValue() == null) {
                             return Result.buildWithParams(false, "Field not equals, name {}, new value is null", f0.getName());
                         } else {
-                            String currentSerializer = UndoLogManager.getCurrentSerializer();
+                            String currentSerializer = AbstractUndoLogManager.getCurrentSerializer();
                             if (StringUtils.equals(currentSerializer, FastjsonUndoLogParser.NAME)) {
                                 convertType(f0, f1);
                             }
