+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/AbstractUndoLogManager.java
@@ -1,0 +1,197 @@
+/*
+ *  Copyright 1999-2019 Seata.io Group.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package io.seata.rm.datasource.undo;
+
+import io.seata.common.util.CollectionUtils;
+import io.seata.config.ConfigurationFactory;
+import io.seata.core.constants.ClientTableColumnsName;
+import io.seata.core.constants.ConfigurationKeys;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+
+/**
+ * @author jsbxyyx
+ * @date 2019/09/07
+ */
+public abstract class AbstractUndoLogManager implements UndoLogManager {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractUndoLogManager.class);
+
+    protected enum State {
+        /**
+         * This state can be properly rolled back by services
+         */
+        Normal(0),
+        /**
+         * This state prevents the branch transaction from inserting undo_log after the global transaction is rolled
+         * back.
+         */
+        GlobalFinished(1);
+
+        private int value;
+
+        State(int value) {
+            this.value = value;
+        }
+
+        public int getValue() {
+            return value;
+        }
+    }
+
+    protected static final String UNDO_LOG_TABLE_NAME = ConfigurationFactory.getInstance()
+            .getConfig(ConfigurationKeys.TRANSACTION_UNDO_LOG_TABLE, ConfigurationKeys.TRANSACTION_UNDO_LOG_DEFAULT_TABLE);
+
+    protected static final String SELECT_UNDO_LOG_SQL = "SELECT * FROM " + UNDO_LOG_TABLE_NAME +
+            " WHERE " + ClientTableColumnsName.UNDO_LOG_BRANCH_XID + " = ? AND " + ClientTableColumnsName.UNDO_LOG_XID + " = ? FOR UPDATE";
+
+    protected static final String DELETE_UNDO_LOG_SQL = "DELETE FROM " + UNDO_LOG_TABLE_NAME +
+            " WHERE " + ClientTableColumnsName.UNDO_LOG_BRANCH_XID + " = ? AND " + ClientTableColumnsName.UNDO_LOG_XID + " = ?";
+
+    private static final ThreadLocal<String> SERIALIZER_LOCAL = new ThreadLocal<>();
+
+    public static String getCurrentSerializer() {
+        return SERIALIZER_LOCAL.get();
+    }
+
+    public static void setCurrentSerializer(String serializer) {
+        SERIALIZER_LOCAL.set(serializer);
+    }
+
+    public static void removeCurrentSerializer() {
+        SERIALIZER_LOCAL.remove();
+    }
+
+    /**
+     * get db type
+     * @return the db type
+     */
+    public abstract String getDbType();
+
+    /**
+     * Delete undo log.
+     *
+     * @param xid      the xid
+     * @param branchId the branch id
+     * @param conn     the conn
+     * @throws SQLException the sql exception
+     */
+    @Override
+    public void deleteUndoLog(String xid, long branchId, Connection conn) throws SQLException {
+        PreparedStatement deletePST = null;
+        try {
+            deletePST = conn.prepareStatement(DELETE_UNDO_LOG_SQL);
+            deletePST.setLong(1, branchId);
+            deletePST.setString(2, xid);
+            deletePST.executeUpdate();
+        } catch (Exception e) {
+            if (!(e instanceof SQLException)) {
+                e = new SQLException(e);
+            }
+            throw (SQLException) e;
+        } finally {
+            if (deletePST != null) {
+                deletePST.close();
+            }
+        }
+    }
+
+    /**
+     * batch Delete undo log.
+     *
+     * @param xids
+     * @param branchIds
+     * @param conn
+     */
+    @Override
+    public void batchDeleteUndoLog(Set<String> xids, Set<Long> branchIds, Connection conn) throws SQLException {
+        if (CollectionUtils.isEmpty(xids) || CollectionUtils.isEmpty(branchIds)) {
+            return;
+        }
+        int xidSize = xids.size();
+        int branchIdSize = branchIds.size();
+        String batchDeleteSql = toBatchDeleteUndoLogSql(xidSize, branchIdSize);
+        PreparedStatement deletePST = null;
+        try {
+            deletePST = conn.prepareStatement(batchDeleteSql);
+            int paramsIndex = 1;
+            for (Long branchId : branchIds) {
+                deletePST.setLong(paramsIndex++,branchId);
+            }
+            for (String xid: xids){
+                deletePST.setString(paramsIndex++, xid);
+            }
+            int deleteRows = deletePST.executeUpdate();
+            if (LOGGER.isDebugEnabled()) {
+                LOGGER.debug("batch delete undo log size " + deleteRows);
+            }
+        }catch (Exception e){
+            if (!(e instanceof SQLException)) {
+                e = new SQLException(e);
+            }
+            throw (SQLException) e;
+        } finally {
+            if (deletePST != null) {
+                deletePST.close();
+            }
+        }
+
+    }
+
+    protected static String toBatchDeleteUndoLogSql(int xidSize, int branchIdSize) {
+        StringBuilder sqlBuilder = new StringBuilder(64);
+        sqlBuilder.append("DELETE FROM ")
+                .append(UNDO_LOG_TABLE_NAME)
+                .append(" WHERE  " + ClientTableColumnsName.UNDO_LOG_BRANCH_XID + " IN ");
+        appendInParam(branchIdSize, sqlBuilder);
+        sqlBuilder.append(" AND " + ClientTableColumnsName.UNDO_LOG_XID + " IN ");
+        appendInParam(xidSize, sqlBuilder);
+        return sqlBuilder.toString();
+    }
+
+    protected static void appendInParam(int size, StringBuilder sqlBuilder) {
+        sqlBuilder.append(" (");
+        for (int i = 0; i < size; i++) {
+            sqlBuilder.append("?");
+            if (i < (size - 1)) {
+                sqlBuilder.append(",");
+            }
+        }
+        sqlBuilder.append(") ");
+    }
+
+    protected static boolean canUndo(int state) {
+        return state == State.Normal.getValue();
+    }
+
+    protected static String buildContext(String serializer) {
+        Map<String, String> map = new HashMap<>();
+        map.put(UndoLogConstants.SERIALIZER_KEY, serializer);
+        return CollectionUtils.encodeMap(map);
+    }
+
+    protected static Map<String, String> parseContext(String data) {
+        return CollectionUtils.decodeMap(data);
+    }
+
+}
