+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/mysql/MySQLUndoLogManager.java
@@ -1,0 +1,312 @@
+/*
+ *  Copyright 1999-2019 Seata.io Group.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package io.seata.rm.datasource.undo.mysql;
+
+import com.alibaba.druid.util.JdbcConstants;
+import io.seata.common.Constants;
+import io.seata.common.exception.NotSupportYetException;
+import io.seata.common.util.BlobUtils;
+import io.seata.core.constants.ClientTableColumnsName;
+import io.seata.core.exception.TransactionException;
+import io.seata.rm.datasource.ConnectionContext;
+import io.seata.rm.datasource.ConnectionProxy;
+import io.seata.rm.datasource.DataSourceProxy;
+import io.seata.rm.datasource.sql.struct.TableMeta;
+import io.seata.rm.datasource.sql.struct.TableMetaCache;
+import io.seata.rm.datasource.undo.AbstractUndoExecutor;
+import io.seata.rm.datasource.undo.AbstractUndoLogManager;
+import io.seata.rm.datasource.undo.BranchUndoLog;
+import io.seata.rm.datasource.undo.SQLUndoLog;
+import io.seata.rm.datasource.undo.UndoExecutorFactory;
+import io.seata.rm.datasource.undo.UndoLogConstants;
+import io.seata.rm.datasource.undo.UndoLogParser;
+import io.seata.rm.datasource.undo.UndoLogParserFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.sql.Blob;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.SQLIntegrityConstraintViolationException;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
+
+import static io.seata.core.exception.TransactionExceptionCode.BranchRollbackFailed_Retriable;
+
+/**
+ * @author jsbxyyx
+ * @date 2019/09/07
+ */
+public class MySQLUndoLogManager extends AbstractUndoLogManager {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(MySQLUndoLogManager.class);
+
+    /**
+     * branch_id, xid, context, rollback_info, log_status, log_created, log_modified
+     */
+    private static final String INSERT_UNDO_LOG_SQL = "INSERT INTO " + UNDO_LOG_TABLE_NAME +
+            " (" + ClientTableColumnsName.UNDO_LOG_BRANCH_XID + ", " + ClientTableColumnsName.UNDO_LOG_XID + ", "
+            + ClientTableColumnsName.UNDO_LOG_CONTEXT + ", " + ClientTableColumnsName.UNDO_LOG_ROLLBACK_INFO + ", "
+            + ClientTableColumnsName.UNDO_LOG_LOG_STATUS + ", " + ClientTableColumnsName.UNDO_LOG_LOG_CREATED + ", "
+            + ClientTableColumnsName.UNDO_LOG_LOG_MODIFIED + ")" +
+            " VALUES (?, ?, ?, ?, ?, now(), now())";
+
+    private static final String DELETE_UNDO_LOG_BY_CREATE_SQL = "DELETE FROM " + UNDO_LOG_TABLE_NAME +
+            " WHERE log_created <= ? LIMIT ?";
+
+    @Override
+    public String getDbType() {
+        return JdbcConstants.MYSQL;
+    }
+
+    /**
+     * Flush undo logs.
+     *
+     * @param cp the cp
+     * @throws SQLException the sql exception
+     */
+    @Override
+    public void flushUndoLogs(ConnectionProxy cp) throws SQLException {
+        assertDbSupport(cp.getDbType());
+
+        ConnectionContext connectionContext = cp.getContext();
+        String xid = connectionContext.getXid();
+        long branchID = connectionContext.getBranchId();
+
+        BranchUndoLog branchUndoLog = new BranchUndoLog();
+        branchUndoLog.setXid(xid);
+        branchUndoLog.setBranchId(branchID);
+        branchUndoLog.setSqlUndoLogs(connectionContext.getUndoItems());
+
+        UndoLogParser parser = UndoLogParserFactory.getInstance();
+        byte[] undoLogContent = parser.encode(branchUndoLog);
+
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Flushing UNDO LOG: {}", new String(undoLogContent, Constants.DEFAULT_CHARSET));
+        }
+
+        insertUndoLogWithNormal(xid, branchID, buildContext(parser.getName()), undoLogContent,
+                cp.getTargetConnection());
+    }
+
+    /**
+     * Undo.
+     *
+     * @param dataSourceProxy the data source proxy
+     * @param xid             the xid
+     * @param branchId        the branch id
+     * @throws TransactionException the transaction exception
+     */
+    @Override
+    public void undo(DataSourceProxy dataSourceProxy, String xid, long branchId) throws TransactionException {
+        assertDbSupport(dataSourceProxy.getDbType());
+
+        Connection conn = null;
+        ResultSet rs = null;
+        PreparedStatement selectPST = null;
+
+        for (; ; ) {
+            try {
+                conn = dataSourceProxy.getPlainConnection();
+
+                // The entire undo process should run in a local transaction.
+                conn.setAutoCommit(false);
+
+                // Find UNDO LOG
+                selectPST = conn.prepareStatement(SELECT_UNDO_LOG_SQL);
+                selectPST.setLong(1, branchId);
+                selectPST.setString(2, xid);
+                rs = selectPST.executeQuery();
+
+                boolean exists = false;
+                while (rs.next()) {
+                    exists = true;
+
+                    // It is possible that the server repeatedly sends a rollback request to roll back
+                    // the same branch transaction to multiple processes,
+                    // ensuring that only the undo_log in the normal state is processed.
+                    int state = rs.getInt(ClientTableColumnsName.UNDO_LOG_LOG_STATUS);
+                    if (!canUndo(state)) {
+                        if (LOGGER.isInfoEnabled()) {
+                            LOGGER.info("xid {} branch {}, ignore {} undo_log",
+                                    xid, branchId, state);
+                        }
+                        return;
+                    }
+
+                    String contextString = rs.getString(ClientTableColumnsName.UNDO_LOG_CONTEXT);
+                    Map<String, String> context = parseContext(contextString);
+                    Blob b = rs.getBlob(ClientTableColumnsName.UNDO_LOG_ROLLBACK_INFO);
+                    byte[] rollbackInfo = BlobUtils.blob2Bytes(b);
+
+                    String serializer = context == null ? null : context.get(UndoLogConstants.SERIALIZER_KEY);
+                    UndoLogParser parser = serializer == null ? UndoLogParserFactory.getInstance() :
+                            UndoLogParserFactory.getInstance(serializer);
+                    BranchUndoLog branchUndoLog = parser.decode(rollbackInfo);
+
+                    try {
+                        // put serializer name to local
+                        setCurrentSerializer(parser.getName());
+                        List<SQLUndoLog> sqlUndoLogs = branchUndoLog.getSqlUndoLogs();
+                        if (sqlUndoLogs.size() > 1) {
+                            Collections.reverse(sqlUndoLogs);
+                        }
+                        for (SQLUndoLog sqlUndoLog : sqlUndoLogs) {
+                            TableMeta tableMeta = TableMetaCache.getTableMeta(dataSourceProxy, sqlUndoLog.getTableName());
+                            sqlUndoLog.setTableMeta(tableMeta);
+                            AbstractUndoExecutor undoExecutor = UndoExecutorFactory.getUndoExecutor(
+                                    dataSourceProxy.getDbType(),
+                                    sqlUndoLog);
+                            undoExecutor.executeOn(conn);
+                        }
+                    } finally {
+                        // remove serializer name
+                        removeCurrentSerializer();
+                    }
+                }
+
+                // If undo_log exists, it means that the branch transaction has completed the first phase,
+                // we can directly roll back and clean the undo_log
+                // Otherwise, it indicates that there is an exception in the branch transaction,
+                // causing undo_log not to be written to the database.
+                // For example, the business processing timeout, the global transaction is the initiator rolls back.
+                // To ensure data consistency, we can insert an undo_log with GlobalFinished state
+                // to prevent the local transaction of the first phase of other programs from being correctly submitted.
+                // See https://github.com/seata/seata/issues/489
+
+                if (exists) {
+                    deleteUndoLog(xid, branchId, conn);
+                    conn.commit();
+                    if (LOGGER.isInfoEnabled()) {
+                        LOGGER.info("xid {} branch {}, undo_log deleted with {}",
+                                xid, branchId, State.GlobalFinished.name());
+                    }
+                } else {
+                    insertUndoLogWithGlobalFinished(xid, branchId, UndoLogParserFactory.getInstance(), conn);
+                    conn.commit();
+                    if (LOGGER.isInfoEnabled()) {
+                        LOGGER.info("xid {} branch {}, undo_log added with {}",
+                                xid, branchId, State.GlobalFinished.name());
+                    }
+                }
+
+                return;
+            } catch (SQLIntegrityConstraintViolationException e) {
+                // Possible undo_log has been inserted into the database by other processes, retrying rollback undo_log
+                if (LOGGER.isInfoEnabled()) {
+                    LOGGER.info("xid {} branch {}, undo_log inserted, retry rollback",
+                            xid, branchId);
+                }
+            } catch (Throwable e) {
+                if (conn != null) {
+                    try {
+                        conn.rollback();
+                    } catch (SQLException rollbackEx) {
+                        LOGGER.warn("Failed to close JDBC resource while undo ... ", rollbackEx);
+                    }
+                }
+                throw new TransactionException(BranchRollbackFailed_Retriable, String.format("%s/%s %s", branchId, xid, e.getMessage()),
+                        e);
+
+            } finally {
+                try {
+                    if (rs != null) {
+                        rs.close();
+                    }
+                    if (selectPST != null) {
+                        selectPST.close();
+                    }
+                    if (conn != null) {
+                        conn.close();
+                    }
+                } catch (SQLException closeEx) {
+                    LOGGER.warn("Failed to close JDBC resource while undo ... ", closeEx);
+                }
+            }
+        }
+    }
+
+    @Override
+    public int deleteUndoLogByLogCreated(Date logCreated, int limitRows, Connection conn) throws SQLException {
+        PreparedStatement deletePST = null;
+        try {
+            deletePST = conn.prepareStatement(DELETE_UNDO_LOG_BY_CREATE_SQL);
+            deletePST.setDate(1, new java.sql.Date(logCreated.getTime()));
+            deletePST.setInt(2, limitRows);
+            int deleteRows = deletePST.executeUpdate();
+            if (LOGGER.isDebugEnabled()) {
+                LOGGER.debug("batch delete undo log size " + deleteRows);
+            }
+            return deleteRows;
+        } catch (Exception e) {
+            if (!(e instanceof SQLException)) {
+                e = new SQLException(e);
+            }
+            throw (SQLException) e;
+        } finally {
+            if (deletePST != null) {
+                deletePST.close();
+            }
+        }
+    }
+
+
+
+    private static void insertUndoLogWithNormal(String xid, long branchID, String rollbackCtx,
+                                                byte[] undoLogContent, Connection conn) throws SQLException {
+        insertUndoLog(xid, branchID, rollbackCtx, undoLogContent, State.Normal, conn);
+    }
+
+    private static void insertUndoLogWithGlobalFinished(String xid, long branchID, UndoLogParser parser,
+                                                        Connection conn) throws SQLException {
+        insertUndoLog(xid, branchID, buildContext(parser.getName()),
+                parser.getDefaultContent(), State.GlobalFinished, conn);
+    }
+
+    private static void insertUndoLog(String xid, long branchID, String rollbackCtx,
+                                      byte[] undoLogContent, State state, Connection conn) throws SQLException {
+        PreparedStatement pst = null;
+        try {
+            pst = conn.prepareStatement(INSERT_UNDO_LOG_SQL);
+            pst.setLong(1, branchID);
+            pst.setString(2, xid);
+            pst.setString(3, rollbackCtx);
+            pst.setBlob(4, BlobUtils.bytes2Blob(undoLogContent));
+            pst.setInt(5, state.getValue());
+            pst.executeUpdate();
+        } catch (Exception e) {
+            if (!(e instanceof SQLException)) {
+                e = new SQLException(e);
+            }
+            throw (SQLException) e;
+        } finally {
+            if (pst != null) {
+                pst.close();
+            }
+        }
+    }
+
+    private static void assertDbSupport(String dbType) {
+        if (!JdbcConstants.MYSQL.equals(dbType)) {
+            throw new NotSupportYetException("DbType[" + dbType + "] is not support yet!");
+        }
+    }
+
+}
